pyproject.toml
[build-system]
requires = ["hatchling"]  # New standard
build-backend = "hatchling.build"

[project]
name = "kundo_movies"
version = "0.0.1"
authors = [
  { name="filip", email="filip.pontus.stenbeck@gmail.com" },
]

dependencies = [  # Få med dependencies (istället för requirment.txt)
    "fastapi[standard]",
    "uvicorn[standard]",
    "requests",
    "jinja2",
    "pytest-asyncio"
]

description = "Kundo movies demo application"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
